requirements.txt
# Core dependencies
openai-whisper>=20231117
click>=8.0
rich>=13.0
pydub>=0.25.0

# Optional: Faster inference (4x speedup on CPU)
# pip install faster-whisper
faster-whisper>=1.0.0

# Optional: OpenAI API backend
# pip install openai
openai>=1.0

# Development
# pip install -r requirements-dev.txt